Changed the way instructions get their register placeholder replaced
In short, I use an Optional<Integer> to point out the position of the argument placeholder, if any. Then, when I need to specify a return register, I call setReturn(), which will replace every placeholder argument in every instruction of the block. Of course, we have to set the argument placeholder to empty() when we replace one, in order for this to work.